FE/Qt: Build warnings.
svn:sync-xref-src-repo-rev: r97853

+++ src/VBox/Frontends/VirtualBox/src/runtime/UIMouseHandler.cpp
@@ -1,4 +1,4 @@
-/* $Id: UIMouseHandler.cpp 53931 2015-01-22 17:24:21Z sergey.dubov@oracle.com $ */
+/* $Id: UIMouseHandler.cpp 53959 2015-01-26 12:02:51Z sergey.dubov@oracle.com $ */
 /** @file
  * VBox Qt GUI - UIMouseHandler class implementation.
  */
@@ -936,8 +936,8 @@ bool UIMouseHandler::mouseEvent(int iEventType, ulong uScreenId,
                 {
                     const double xScaleFactor = (double)scaledSize.width()  / pFrameBuffer->width();
                     const double yScaleFactor = (double)scaledSize.height() / pFrameBuffer->height();
-                    cpnt.setX(cpnt.x() / xScaleFactor);
-                    cpnt.setY(cpnt.y() / yScaleFactor);
+                    cpnt.setX((int)(cpnt.x() / xScaleFactor));
+                    cpnt.setY((int)(cpnt.y() / yScaleFactor));
                 }
             }
 
@@ -948,8 +948,8 @@ bool UIMouseHandler::mouseEvent(int iEventType, ulong uScreenId,
                 const double dBackingScaleFactor = darwinBackingScaleFactor(m_windows.value(uScreenId));
                 if (dBackingScaleFactor > 1.0)
                 {
-                    cpnt.setX(cpnt.x() * dBackingScaleFactor);
-                    cpnt.setY(cpnt.y() * dBackingScaleFactor);
+                    cpnt.setX((int)(cpnt.x() * dBackingScaleFactor));
+                    cpnt.setY((int)(cpnt.y() * dBackingScaleFactor));
                 }
             }
 #endif /* Q_WS_MAC */
